requirements.txt
ultralytics==8.1.34
streamlit==1.35.0
opencv-python==4.11.0.1
numpy==1.24.4
pandas==2.2.2
matplotlib==3.8.4
Pillow==10.3.0
requests==2.31.0
torch==2.3.0
torchvision==0.18.0
scikit-learn==1.4.2
deep_sort_realtime==1.3.1
python-dotenv==1.0.1
